requirements.txt
djangorestframework
markdown 
python-decouple
django-filter  
qrcode
requests
certifi
pyopenssl